Planet mesh and collider changed, misc
Two new meshes add via Blender: UV sphere and icosphere, applied to
planet and largest moon, respectively. Both are low-poly, as mesh
colliders only work for objects with <255 faces. This solves the issue
of jittering, which was caused by the smoothness of the sphere collider
(meaning that each object only touched the ground at a single point). A
new issue is that non-convex mesh colliders can't collide with each
other (though they can collide with primitive colliders), so either we
have to ensure that planets never collide and the larger parts of
smaller bodies are primitive colliders, or we can make planets convex,
which will prevent us from making valleys.

Additionally:
* sun intensity lowered
* new planet texture, should be mapped properly
* mass calculation is done via renderer bounds instead of collider
bounds, which is necessary in order to use mesh colliders
* Issue: planet textures are really low-res, we'll have to make our own