FajnerCube cannot into space
I fixed the bug where the FajnerCube could jump again in the air,
seeming as if he could fly. This is no longer. I solved it be adding
another condition to the jump if statement. This condition is a
raytrace downwards until it hits an object, and it must be within the
height (+0.1).